+++ packages/rum-core/src/boot/rum.ts
@@ -1,10 +1,11 @@
 import { combine, commonInit, Configuration } from '@datadog/browser-core'
-import { startDOMMutationCollection } from '../browser/domMutationCollection'
+import { createDOMMutationObservable } from '../browser/domMutationObservable'
 import { startPerformanceCollection } from '../browser/performanceCollection'
 import { startRumAssembly } from '../domain/assembly'
 import { startInternalContext } from '../domain/internalContext'
 import { LifeCycle } from '../domain/lifeCycle'
 import { startParentContexts } from '../domain/parentContexts'
+import { startForegroundContexts } from '../domain/foregroundContexts'
 import { startRequestCollection } from '../domain/requestCollection'
 import { startActionCollection } from '../domain/rumEventsCollection/action/actionCollection'
 import { startErrorCollection } from '../domain/rumEventsCollection/error/errorCollection'
@@ -20,9 +21,9 @@ import { RumUserConfiguration } from './rumPublicApi'
 
 export function startRum(userConfiguration: RumUserConfiguration, getCommonContext: () => CommonContext) {
   const lifeCycle = new LifeCycle()
-
   const { configuration, internalMonitoring } = commonInit(userConfiguration, buildEnv)
   const session = startRumSession(configuration, lifeCycle)
+  const domMutationObservable = createDOMMutationObservable()
 
   internalMonitoring.setExternalContextProvider(() =>
     combine(
@@ -34,7 +35,7 @@ export function startRum(userConfiguration: RumUserConfiguration, getCommonConte
     )
   )
 
-  const { parentContexts } = startRumEventCollection(
+  const { parentContexts, foregroundContexts } = startRumEventCollection(
     userConfiguration.applicationId,
     lifeCycle,
     configuration,
@@ -44,13 +45,13 @@ export function startRum(userConfiguration: RumUserConfiguration, getCommonConte
 
   startLongTaskCollection(lifeCycle)
   startResourceCollection(lifeCycle, session)
-  const { addTiming, startView } = startViewCollection(lifeCycle, location)
-  const { addError } = startErrorCollection(lifeCycle, configuration)
-  const { addAction } = startActionCollection(lifeCycle, configuration)
+
+  const { addTiming, startView } = startViewCollection(lifeCycle, location, domMutationObservable, foregroundContexts)
+  const { addError } = startErrorCollection(lifeCycle, configuration, foregroundContexts)
+  const { addAction } = startActionCollection(lifeCycle, domMutationObservable, configuration, foregroundContexts)
 
   startRequestCollection(lifeCycle, configuration)
   startPerformanceCollection(lifeCycle, configuration)
-  startDOMMutationCollection(lifeCycle)
 
   const internalContext = startInternalContext(userConfiguration.applicationId, session, parentContexts)
 
@@ -75,15 +76,20 @@ export function startRumEventCollection(
   getCommonContext: () => CommonContext
 ) {
   const parentContexts = startParentContexts(lifeCycle, session)
+  const foregroundContexts = startForegroundContexts(configuration)
   const batch = startRumBatch(configuration, lifeCycle)
+
   startRumAssembly(applicationId, configuration, lifeCycle, session, parentContexts, getCommonContext)
 
   return {
     parentContexts,
+    foregroundContexts,
     stop: () => {
       // prevent batch from previous tests to keep running and send unwanted requests
       // could be replaced by stopping all the component when they will all have a stop method
       batch.stop()
+      parentContexts.stop()
+      foregroundContexts.stop()
     },
   }
 }