+++ packages/rum-core/src/domain/assembly.ts
@@ -51,6 +51,8 @@ const OTHER_EVENTS_MODIFIABLE_FIELD_PATHS = [
   'context',
 ]
 
+type Mutable<T> = { -readonly [P in keyof T]: T[P] }
+
 export function startRumAssembly(
   applicationId: string,
   configuration: Configuration,
@@ -93,12 +95,11 @@ export function startRumAssembly(
         serverRumEvent.context = combine(commonContext.context, customerContext)
 
         if (!('has_replay' in serverRumEvent.session)) {
-          ;(serverRumEvent.session as { has_replay?: boolean }).has_replay = commonContext.hasReplay
+          ;(serverRumEvent.session as Mutable<RumEvent['session']>).has_replay = commonContext.hasReplay
         }
 
         if (!isEmptyObject(commonContext.user)) {
-          // eslint-disable-next-line @typescript-eslint/no-unnecessary-type-assertion
-          ;(serverRumEvent.usr as RumEvent['usr']) = commonContext.user as User & Context
+          ;(serverRumEvent.usr as Mutable<RumEvent['usr']>) = commonContext.user as User & Context
         }
         if (shouldSend(serverRumEvent, configuration.beforeSend, errorFilter)) {
           if (isEmptyObject(serverRumEvent.context)) {