+++ packages/rum-core/src/domain/rumEventsCollection/resource/resourceUtils.ts
@@ -19,6 +19,7 @@ export interface PerformanceResourceDetails {
   dns?: PerformanceResourceDetailsElement
   connect?: PerformanceResourceDetailsElement
   ssl?: PerformanceResourceDetailsElement
+  // eslint-disable-next-line camelcase
   first_byte: PerformanceResourceDetailsElement
   download: PerformanceResourceDetailsElement
 }