+++ packages/rum-core/src/domain/trackPageActivities.spec.ts
@@ -34,16 +34,23 @@ function eventsCollector<T>() {
 
 describe('trackPagePageActivities', () => {
   const { events, pushEvent } = eventsCollector<PageActivityEvent>()
+
+  let lifeCycle: LifeCycle
+  let domMutationObservable: Observable<void>
+
+  beforeEach(() => {
+    lifeCycle = new LifeCycle()
+    domMutationObservable = new Observable()
+  })
+
   it('emits an activity event on dom mutation', () => {
-    const lifeCycle = new LifeCycle()
-    trackPageActivities(lifeCycle).observable.subscribe(pushEvent)
-    lifeCycle.notify(LifeCycleEventType.DOM_MUTATED)
+    trackPageActivities(lifeCycle, domMutationObservable).observable.subscribe(pushEvent)
+    domMutationObservable.notify()
     expect(events).toEqual([{ isBusy: false }])
   })
 
   it('emits an activity event on resource collected', () => {
-    const lifeCycle = new LifeCycle()
-    trackPageActivities(lifeCycle).observable.subscribe(pushEvent)
+    trackPageActivities(lifeCycle, domMutationObservable).observable.subscribe(pushEvent)
     const performanceTiming = {
       entryType: 'resource',
     }
@@ -52,8 +59,7 @@ describe('trackPagePageActivities', () => {
   })
 
   it('does not emit an activity event when a navigation occurs', () => {
-    const lifeCycle = new LifeCycle()
-    trackPageActivities(lifeCycle).observable.subscribe(pushEvent)
+    trackPageActivities(lifeCycle, domMutationObservable).observable.subscribe(pushEvent)
     const performanceTiming = {
       entryType: 'navigation',
     }
@@ -65,17 +71,16 @@ describe('trackPagePageActivities', () => {
   })
 
   it('stops emitting activities after calling stop()', () => {
-    const lifeCycle = new LifeCycle()
-    const { stop, observable } = trackPageActivities(lifeCycle)
+    const { stop, observable } = trackPageActivities(lifeCycle, domMutationObservable)
     observable.subscribe(pushEvent)
 
-    lifeCycle.notify(LifeCycleEventType.DOM_MUTATED)
+    domMutationObservable.notify()
     expect(events).toEqual([{ isBusy: false }])
 
     stop()
 
-    lifeCycle.notify(LifeCycleEventType.DOM_MUTATED)
-    lifeCycle.notify(LifeCycleEventType.DOM_MUTATED)
+    domMutationObservable.notify()
+    domMutationObservable.notify()
 
     expect(events).toEqual([{ isBusy: false }])
   })
@@ -84,9 +89,16 @@ describe('trackPagePageActivities', () => {
     function makeFakeRequestCompleteEvent(requestIndex: number) {
       return { requestIndex } as RequestCompleteEvent
     }
+    let lifeCycle: LifeCycle
+    let domMutationObservable: Observable<void>
+
+    beforeEach(() => {
+      lifeCycle = new LifeCycle()
+      domMutationObservable = new Observable()
+    })
+
     it('emits an activity event when a request starts', () => {
-      const lifeCycle = new LifeCycle()
-      trackPageActivities(lifeCycle).observable.subscribe(pushEvent)
+      trackPageActivities(lifeCycle, domMutationObservable).observable.subscribe(pushEvent)
       lifeCycle.notify(LifeCycleEventType.REQUEST_STARTED, {
         requestIndex: 10,
       })
@@ -95,7 +107,7 @@ describe('trackPagePageActivities', () => {
 
     it('emits an activity event when a request completes', () => {
       const lifeCycle = new LifeCycle()
-      trackPageActivities(lifeCycle).observable.subscribe(pushEvent)
+      trackPageActivities(lifeCycle, domMutationObservable).observable.subscribe(pushEvent)
       lifeCycle.notify(LifeCycleEventType.REQUEST_STARTED, {
         requestIndex: 10,
       })
@@ -105,14 +117,14 @@ describe('trackPagePageActivities', () => {
 
     it('ignores requests that has started before', () => {
       const lifeCycle = new LifeCycle()
-      trackPageActivities(lifeCycle).observable.subscribe(pushEvent)
+      trackPageActivities(lifeCycle, domMutationObservable).observable.subscribe(pushEvent)
       lifeCycle.notify(LifeCycleEventType.REQUEST_COMPLETED, makeFakeRequestCompleteEvent(10))
       expect(events).toEqual([])
     })
 
     it('keeps emitting busy events while all requests are not completed', () => {
       const lifeCycle = new LifeCycle()
-      trackPageActivities(lifeCycle).observable.subscribe(pushEvent)
+      trackPageActivities(lifeCycle, domMutationObservable).observable.subscribe(pushEvent)
       lifeCycle.notify(LifeCycleEventType.REQUEST_STARTED, {
         requestIndex: 10,
       })