+++ packages/rum-core/test/specHelper.ts
@@ -5,11 +5,14 @@ import {
   Configuration,
   Context,
   DEFAULT_CONFIGURATION,
+  Observable,
   TimeStamp,
+  noop,
 } from '@datadog/browser-core'
 import { SPEC_ENDPOINTS, mockClock, Clock } from '../../core/test/specHelper'
 import { LifeCycle, LifeCycleEventType } from '../src/domain/lifeCycle'
 import { ParentContexts } from '../src/domain/parentContexts'
+import { ForegroundContexts } from '../src/domain/foregroundContexts'
 import { RumSession } from '../src/domain/rumSession'
 import { CommonContext, RawRumEvent, RumContext, ViewContext } from '../src/rawRumEvent.types'
 import { validateFormat } from './formatValidation'
@@ -19,6 +22,7 @@ export interface TestSetupBuilder {
   withSession: (session: RumSession) => TestSetupBuilder
   withConfiguration: (overrides: Partial<Configuration>) => TestSetupBuilder
   withParentContexts: (stub: Partial<ParentContexts>) => TestSetupBuilder
+  withForegroundContexts: (stub: Partial<ForegroundContexts>) => TestSetupBuilder
   withFakeClock: () => TestSetupBuilder
   beforeBuild: (callback: BeforeBuildCallback) => TestSetupBuilder
 
@@ -29,15 +33,18 @@ export interface TestSetupBuilder {
 type BeforeBuildCallback = (buildContext: BuildContext) => void | { stop: () => void }
 interface BuildContext {
   lifeCycle: LifeCycle
+  domMutationObservable: Observable<void>
   configuration: Readonly<Configuration>
   session: RumSession
   location: Location
   applicationId: string
   parentContexts: ParentContexts
+  foregroundContexts: ForegroundContexts
 }
 
 export interface TestIO {
   lifeCycle: LifeCycle
+  domMutationObservable: Observable<void>
   clock: Clock
   fakeLocation: Partial<Location>
   session: RumSession
@@ -56,6 +63,7 @@ export function setup(): TestSetupBuilder {
     isTrackedWithResource: () => true,
   }
   const lifeCycle = new LifeCycle()
+  const domMutationObservable = new Observable<void>()
   const cleanupTasks: Array<() => void> = []
   const beforeBuildTasks: BeforeBuildCallback[] = []
   const rawRumEvents: Array<{
@@ -68,6 +76,11 @@ export function setup(): TestSetupBuilder {
   let clock: Clock
   let fakeLocation: Partial<Location> = location
   let parentContexts: ParentContexts
+  let foregroundContexts: ForegroundContexts = {
+    getInForeground: () => undefined,
+    getInForegroundPeriods: () => undefined,
+    stop: noop,
+  }
   const configuration: Partial<Configuration> = {
     ...DEFAULT_CONFIGURATION,
     ...SPEC_ENDPOINTS,
@@ -113,6 +126,10 @@ export function setup(): TestSetupBuilder {
       parentContexts = stub as ParentContexts
       return setupBuilder
     },
+    withForegroundContexts(stub: Partial<ForegroundContexts>) {
+      foregroundContexts = { ...foregroundContexts, ...stub }
+      return setupBuilder
+    },
     withFakeClock() {
       clock = mockClock()
       return setupBuilder
@@ -125,7 +142,9 @@ export function setup(): TestSetupBuilder {
       beforeBuildTasks.forEach((task) => {
         const result = task({
           lifeCycle,
+          domMutationObservable,
           parentContexts,
+          foregroundContexts,
           session,
           applicationId: FAKE_APP_ID,
           configuration: configuration as Configuration,
@@ -139,6 +158,7 @@ export function setup(): TestSetupBuilder {
         clock,
         fakeLocation,
         lifeCycle,
+        domMutationObservable,
         rawRumEvents,
         session,
       }