+++ test/e2e/lib/helpers/validation.ts
@@ -4,7 +4,7 @@ import rumEventsFormatJson from '../../../../rum-events-format/rum-events-format
 import _commonSchemaJson from '../../../../rum-events-format/schemas/_common-schema.json'
 import actionSchemaJson from '../../../../rum-events-format/schemas/action-schema.json'
 import errorSchemaJson from '../../../../rum-events-format/schemas/error-schema.json'
-import long_taskSchemaJson from '../../../../rum-events-format/schemas/long_task-schema.json'
+import longTaskSchemaJson from '../../../../rum-events-format/schemas/long_task-schema.json'
 import resourceSchemaJson from '../../../../rum-events-format/schemas/resource-schema.json'
 import viewSchemaJson from '../../../../rum-events-format/schemas/view-schema.json'
 
@@ -18,7 +18,7 @@ export function validateFormat(events: RumEvent[]) {
       .addSchema(viewSchemaJson, 'schemas/view-schema.json')
       .addSchema(actionSchemaJson, 'schemas/action-schema.json')
       .addSchema(resourceSchemaJson, 'schemas/resource-schema.json')
-      .addSchema(long_taskSchemaJson, 'schemas/long_task-schema.json')
+      .addSchema(longTaskSchemaJson, 'schemas/long_task-schema.json')
       .addSchema(errorSchemaJson, 'schemas/error-schema.json')
       .addSchema(rumEventsFormatJson, 'rum-events-format.json')
       .validate('rum-events-format.json', event)